Menu Functionality 2
Finished implementation of GameScreen.cs, the abstract base screen
class.
Added StartScreen.cs, the main menu for the game.
Added SplashScreen.cs, a potential loading screen.
Added ActionScreen.cs, the main gameplay screen.
Added PauseScreen.cs, the pause menu for the game.
Updated Game1.cs for the implementation of the new screens.
Changed one of the actions in the InputManager to suit the
implementation of the screens.
Added comments in some of the classes.