Tile Engine 2
Moved all of the screen management classes into the GameScreen.cs
directory.
Updated several classes with comments.
Updated the tile engine to read tile data from a file.  Externally, tile
maps will be created with an editor called Tiled, which exports files in
several formats.  The tile engine has been equipped with the ability to
read the text file format that Tiled generates.
Updated the tile engine to deal with multiple layers.
Added Camera.cs, a camera class that adds scrolling to the game.
Added Level.cs, a class that manages the content drawn to the screen for
each level.

Notes: In the future, Level.cs will act as a base class for Level
classes that will inherit from it.  These Level classes will be
responsible for loading all content in each level and managing level
specific enemies and events.