Animation 2 / Camera Update
Created AnimatedSprite.cs, a class that all sprites will inherit from.
Added two camera modes, 'follow' and 'free'.  In 'follow', the camera
will follow the sprite, and in 'free' the player can freely control the
camera independent of the player character.
Added more keys to the Input Manager for player control.
Added Direction.cs, a class that provides enumerations for storing
sprite orientation for rotating the sprites based on the direction
they're moving.