Tile Engine 4 / Collision 1 / SpriteManager
Updated Level.cs with the ability to read larger tilesets (with more
than 9 tiles).  It will also read a file with collision information and
set the collision type to the indicated tile.  A function,
SpawnSprites(), was also added.  It reads in information about the enemy
type and starting position for the level and creates them.
Updated AnimatedSprite.cs.  Sprites will now rotate to face the
direction they're currently moving.
Tile.cs and MapLayer.cs were update to support tile collsions.
Circle.cs is a new shape class that can create a bounding circle for an
object for collision checking.
Collision.cs is the new collision class.  It will hold methods used for
checking collisions and determining the results of them.
A Sprite Manager was added.  It controls the creation, updating, and
drawing of all sprites in the game.
The shells for the new Enemy classes were created (Enemy.cs and it's
inherited Turret.cs).
Note:  While the logic is there, the new collision and spawn reading
implemented in Level.cs is not yet running.  Collision logic has not
been implemented, and the SpriteManager has not been added to the
Initialize and Update calls of Game1.cs.