ProjectileManager2 / DialogueScreen
Added a new property to projectiles, ProjectileAlignment.  It keeps
track of whether a friendly or a foe fired the laser.  The projectiles
alignment will be used for projectile to sprite collision tracking
later.
Added DialogueScreen.cs, a conversation window.  Informational text
appears that the player can cycle through.
Updated the file organization of the classes.  Added misc. comments.